requirements.txt
selenium
mysql-connector-python
python-dotenv
requests
